requirements.txt
pandas == 2.2.3
streamlit == 1.51.0
fastapi>=0.110
uvicorn>=0.29
python-dotenv>=1.0
httpx>=0.27
langfuse>=2.33
langgraph>=0.0.48
langchain-openai>=0.1.10
langchain-core>=0.1.52
langchain==0.3.27
tavily-python == 0.7.12
numpy>=1.26
guardrails-ai
mcp[cli]>=0.1